Laid out cmake files
