+++ .github/workflows/emsdk.Dockerfile
@@ -24,8 +24,16 @@ WORKDIR /home/Downloads
 # Install qpdf and create non root user
 # The packages libjsoncpp-dev libjsoncpp-doc libjsoncpp1 are for JSON reading
 # The libwebp-dev package is for saving images to webp
+# For latest GCC installing gcc-12 gcc-12-base gcc-12-doc g++-12
+# For latest G++ installing libstdc++-12-dev libstdc++-12-doc 
+# Installing cmake-curses-gui to use ccmake
 RUN sudo apt-get update &&          \
-    sudo apt-get -y install build-essential gdb cmake zlib1g-dev libjpeg-dev libgnutls28-dev libssl-dev libjsoncpp-dev libjsoncpp-doc libtiff-tools libwebp-dev qpdf libqpdf-dev ghostscript; \
+    sudo apt-get -y install build-essential gdb cmake   \
+    zlib1g-dev libjpeg-dev libgnutls28-dev libssl-dev   \
+    libjsoncpp-dev libjsoncpp-doc libtiff-tools         \
+    libwebp-dev qpdf libqpdf-dev ghostscript gcc-10     \
+    gcc-12-base gcc-12-doc g++-12 libstdc++-12-dev      \
+    libstdc++-12-doc clang-12 cmake-curses-gui;                          \
 # User created in original container
 # Can be found in source code here https://github.com/trzecieu/emscripten-docker/blob/master/docker/trzeci/emscripten-slim/Dockerfile#L352
     groupmod -g "${GID}" emscripten && usermod -u "${UID}" -g "${GID}" emscripten
