+++ .gitignore
@@ -14,3 +14,6 @@ node_modules
 
 # python virtual environment
 venv
+
+# CMake User presets
+CMakeUserPresets.json
