+++ pdf-to-webp/CMakeLists.txt
@@ -1,12 +1,9 @@
+# TODO: look at https://stunlock.gg/posts/emscripten_with_cmake/
 cmake_minimum_required(VERSION 3.21.0)
-project(pdf_to_webp VERSION 0.1.0)
-
-include(CTest)
-enable_testing()
-
-set(SOURCES main.cpp shared_array.cpp)
-add_executable(pdf_to_webp ${SOURCES})
-
-set(CPACK_PROJECT_NAME ${PROJECT_NAME})
-set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
-include(CPack)
+project(
+    pdf_to_webp
+    VERSION 0.0.1
+    LANGUAGES CXX
+)
+add_subdirectory(components)
+add_subdirectory(app)
