+++ pdf-to-webp/CMakePresets.json
@@ -1,0 +1,8 @@
+{
+    "version": 3,
+    "cmakeMinimumRequired": {
+        "major": 3,
+        "minor": 21,
+        "patch": 0
+    }
+}
