--- pdf-to-webp/artwork_handler.cpp
@@ -1,14 +1,0 @@
-
-// JSON streaming
-#include <jsoncpp/json/json.h>
-
-// WEBP image storing
-#include <webp/decode.h>
-#include <webp/encode.h>
-#include <webp/types.h>
-
-#include "artwork_handler.h"
-
-using namespace pdftowebp;
-
-// TODO: All the fancy image making
