--- pdf-to-webp/artwork_handler.h
@@ -1,7 +1,0 @@
-#ifndef ARTWORK_HANDLER
-#define ARTWORK_HANDLER
-
-namespace pdftowebp {
-
-}
-#endif
