+++ pdf-to-webp/components/artwork_handler/artwork_handler.cpp
@@ -1,0 +1,14 @@
+
+// JSON streaming
+#include <jsoncpp/json/json.h>
+
+// WEBP image storing
+#include <webp/decode.h>
+#include <webp/encode.h>
+#include <webp/types.h>
+
+#include "artwork_handler.h"
+
+using namespace libpdftowebp;
+
+// TODO: All the fancy image making
