+++ pdf-to-webp/components/artwork_handler/artwork_handler.h
@@ -1,0 +1,7 @@
+#ifndef ARTWORK_HANDLER
+#define ARTWORK_HANDLER
+
+namespace libpdftowebp {
+
+}
+#endif
